- Moved action_inputs.py file from the 1st regime root dir. This logic is for more regimes.
